+++ Source/Objects/Hardware/Vme/Bit3 Adaptor/ORBit3Model.m
@@ -78,7 +78,7 @@
 }
 
 -(void) readLongBlock:(unsigned long *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberLongs
 		   withAddMod:(unsigned short) anAddressModifier
 					   usingAddSpace:(unsigned short) anAddressSpace
@@ -95,7 +95,7 @@
 }
 
 -(void) writeLongBlock:(unsigned long *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberLongs
 			withAddMod:(unsigned short) anAddressModifier
 						   usingAddSpace:(unsigned short) anAddressSpace
@@ -113,7 +113,7 @@
 }
 
 -(void) readLong:(unsigned long *) readAddress
-	   atAddress:(unsigned int) vmeAddress
+	   atAddress:(unsigned long) vmeAddress
 	 timesToRead:(unsigned int) numberLongs
 	  withAddMod:(unsigned short) anAddressModifier
    usingAddSpace:(unsigned short) anAddressSpace
@@ -130,7 +130,7 @@
 }
 
 -(void) readByteBlock:(unsigned char *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberBytes
 		   withAddMod:(unsigned short) anAddressModifier
 					   usingAddSpace:(unsigned short) anAddressSpace
@@ -148,7 +148,7 @@
 }
 
 -(void) writeByteBlock:(unsigned char *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberBytes
 			withAddMod:(unsigned short) anAddressModifier
 						   usingAddSpace:(unsigned short) anAddressSpace
@@ -167,7 +167,7 @@
 
 
 -(void) readWordBlock:(unsigned short *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberWords
 		   withAddMod:(unsigned short) anAddressModifier
 					   usingAddSpace:(unsigned short) anAddressSpace
@@ -185,7 +185,7 @@
 }
 
 -(void) writeWordBlock:(unsigned short *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberWords
 			withAddMod:(unsigned short) anAddressModifier
 						   usingAddSpace:(unsigned short) anAddressSpace
