+++ Source/Objects/Hardware/Vme/VMEcpu/ORVmecpuModel.m
@@ -326,7 +326,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
 }
 
 -(void) readLongBlock:(unsigned long *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberLongs
 		   withAddMod:(unsigned short) anAddressModifier
 		usingAddSpace:(unsigned short) anAddressSpace
@@ -342,7 +342,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
 }
 
 -(void) writeLongBlock:(unsigned long *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberLongs
 			withAddMod:(unsigned short) anAddressModifier
 		 usingAddSpace:(unsigned short) anAddressSpace
@@ -363,7 +363,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
  auto-increment.  We use a specific address space to
  do this: 0xFF. */
 -(void) readLong:(unsigned long *) readAddress
-	   atAddress:(unsigned int) vmeAddress
+	   atAddress:(unsigned long) vmeAddress
 	 timesToRead:(unsigned int) numberLongs
 	  withAddMod:(unsigned short) anAddressModifier
    usingAddSpace:(unsigned short) anAddressSpace
@@ -379,7 +379,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
 }
 
 -(void) readByteBlock:(unsigned char *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberBytes
 		   withAddMod:(unsigned short) anAddressModifier
 		usingAddSpace:(unsigned short) anAddressSpace
@@ -396,7 +396,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
 }
 
 -(void) writeByteBlock:(unsigned char *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberBytes
 			withAddMod:(unsigned short) anAddressModifier
 		 usingAddSpace:(unsigned short) anAddressSpace
@@ -413,7 +413,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
 }
 
 -(void) readWordBlock:(unsigned short *) readAddress
-			atAddress:(unsigned int) vmeAddress
+			atAddress:(unsigned long) vmeAddress
 			numToRead:(unsigned int) numberWords
 		   withAddMod:(unsigned short) anAddressModifier
 		usingAddSpace:(unsigned short) anAddressSpace
@@ -431,7 +431,7 @@ NSString* ORVmecpuLock = @"ORVmecpuLock";
 }
 
 -(void) writeWordBlock:(unsigned short *) writeAddress
-			 atAddress:(unsigned int) vmeAddress
+			 atAddress:(unsigned long) vmeAddress
 			numToWrite:(unsigned int) numberWords
 			withAddMod:(unsigned short) anAddressModifier
 		 usingAddSpace:(unsigned short) anAddressSpace
